Pre-alpha upload
Got the basics of the app down, tho I still need to implement the core of the app; the geo-tracking. It sounds kinda sketchy, but the idea is that most phones are already pinging location every few minutes, we can just check if they're in a store and give them a list notification.